WIK-57: Pages de challenge - Manage file name and organize url - Fix players game settings to 10 by default
