+++ src/route.tsx
@@ -4,7 +4,7 @@ import WaitingRoom from "./pages/WaitingRoom/WaitingRoom.tsx";
 import Game from "./pages/Game/Game.tsx";
 import Result from "./pages/Result/Result.tsx";
 import GameChallenge from "./pages/Challenge/Game/Game.tsx";
-import ResultChallenge from "./pages/Challenge/Result/Result.tsx";
+import Challenge from "./pages/Challenge/Challenge.tsx";
 
 const AppRoutes = () => {
     return (
@@ -14,8 +14,8 @@ const AppRoutes = () => {
                 <Route path="/room" element={<WaitingRoom />} />
                 <Route path="/game" element={<Game />} />
                 <Route path="/result" element={<Result />} />
+                <Route path="/challenge" element={<Challenge />} />
                 <Route path="/challenge/game" element={<GameChallenge />} />
-                <Route path="/challenge/result" element={<ResultChallenge />} />
             </Routes>
         </Router>
     );
